requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
jinja2

# RAG + Qdrant + 임베딩
langchain
langchain-community
langchain-qdrant
langchain-text-splitters
sentence-transformers
qdrant-client
